chore: Update modal-root element in layout.js and Cookies key in register/page.js
Update the modal-root element in layout.js to use self-closing syntax. Also, update the key used for setting the token in the Cookies object in register/page.js to "accessToken" instead of "token".

Note: This message assumes that the repository's commit message conventions follow the format of "type: commit message".